phase_3/requirements.txt
# Phase 3: stdlib + optional python-dotenv for local .env loading.
python-dotenv>=1.0.0

